Refactor and comment on accrual service
